docs(transaction): #881 Phase E — YARD, reference page, spec reorganisation (#887)
The cache contract becomes discoverable. YARD on the memoised methods
and cache-affecting setters links to the new reference page; the
reference page is the single source of truth for the invalidation
table and escape-hatch usage.

- YARD @note on memoised methods (Tx#to_binary, Tx#wtxid, the three
  hash_* component methods, TransactionInput#outpoint_binary/
  to_binary, TransactionOutput#to_binary) pointing at the reference
  page
- YARD @!attribute [rw] blocks on the four cache-affecting setters
  (input.sequence=, input.unlocking_script=, output.satoshis=,
  output.locking_script=) — invalidation cost called out where the
  caller pays it
- YARD @raise on Tx#add_input / add_output with @see anchor
- YARD @example on Tx#invalidate_caches showing the only pattern
  that needs it
- Class-level YARD warning on Tx: not thread-safe; direct array
  mutation bypasses invalidation
- New docs/reference/sighash-cache.md with Russian-doll diagram,
  invalidation contract table, one-owner invariant, escape hatch,
  build->sign/verify idiom, thread-safety note. Anchors #one-owner,
  #escape-hatch, #invalidation-contract match the YARD @see clauses
- README cross-link to the new reference page
- docs/reference/index.md updated with explicit entry
- Spec reorganisation: invalidation contract rows live in
  tx_cache_invalidation_spec.rb with parameterised shared examples;
  lifecycle specs (backref, dup, stubs) in tx_cache_lifecycle_spec.rb;
  tx_cache_spec.rb emptied to a breadcrumb comment
- Anchor stability smoke spec prevents YARD/docs drift
