+++ README.md
@@ -125,6 +125,7 @@ Full documentation is available at **[sgbett.github.io/bsv-ruby-sdk](https://sgb
 **Additional resources:**
 
 - [API Reference](https://sgbett.github.io/bsv-ruby-sdk/reference/api/) — auto-generated from YARD annotations
+- [Sighash & Wire Cache](https://sgbett.github.io/bsv-ruby-sdk/reference/sighash-cache/) — Russian-doll memo cache, invalidation contract, escape hatch
 - [spec/ directory](https://github.com/sgbett/bsv-ruby-sdk/tree/master/spec) — runnable usage examples
 - Changelogs: [sdk](gem/bsv-sdk/CHANGELOG.md) · [attest](gem/bsv-attest/CHANGELOG.md)
