+++ docs/reference/index.md
@@ -12,4 +12,5 @@ Canonical reference documents and auto-generated API documentation.
 - [wtxid / dtxid Convention](wtxid-dtxid.md) — wire-order vs display-order transaction IDs
 - [Conformance Vectors](conformance-vectors.md) — cross-SDK test vector validation
 - [Wycheproof Malleability Analysis](wycheproof-malleability-analysis.md) — ECDSA vector analysis
+- [Sighash & Wire Cache](sighash-cache.md) — Russian-doll memo cache, invalidation contract, escape hatch
 - [API Reference](api/) — auto-generated from YARD annotations
